// web/src/domain/types/alert.ts
export enum AlertActionTypes {
  DISPATCH_ALERT_MESSAGE = 'DISPATCH_ALERT_MESSAGE',
  DISPATCH_ALERT_SUCCESS = 'DISPATCH_ALERT_SUCCESS',
  DISPATCH_ALERT_FAILURE = 'DISPATCH_ALERT_FAILURE',
  DISPATCH_ALERT_ERROR = 'DISPATCH_ALERT_ERROR',
  DISPATCH_ALERT_CRITICAL = 'DISPATCH_ALERT_CRITICAL',
  DISPATCH_ALERT_NONCRITICAL = 'DISPATCH_ALERT_NONCRITICAL',
  DISMISS_SELECTED_ALERT = 'DISMISS_SELECTED_ALERT',
  DISMISS_ALL_ALERTS = 'DISMISS_ALL_ALERTS',
}